Detects faces, more accuracy(almost correctly)
There is more accuracy with the detection where the right paddle does deflect the balls (for the most part). There is still an issue with the actual boundry being too far away from the paddle, which does need fixing. However, from very limited testing, if it is in position, it will deflect the ball.